⚡️ 퍼포먼스 관련 script수정 kakao script를 사용 하는 모든페이지에서 불러오도록 코딩
